Add initial commit of Rockchip kernel sources.
This is a patch form of an unpacked, historyless tarball from:
https://github.com/Galland/rk3x_kernel_3.0.36.git

Although this is Android source, I would like to have the proper
Linux history as well for forward-porting RK29 support to a
non-ancient kernel version.

This change is a diff of the v3.0.36 tag with the initial Curie
source (9d93fa758939738012ba353ceb033e812e56c838).